requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
pdfplumber>=0.10.0
chromadb>=0.4.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
xlrd>=2.0.0
